trivial: split out StoragePricing
