+++ aptos-move/aptos-abstract-gas-usage/src/algebra.rs
@@ -6,7 +6,7 @@ use aptos_gas_algebra::{
 };
 use aptos_gas_meter::GasAlgebra;
 use aptos_gas_schedule::VMGasParameters;
-use aptos_vm_types::storage::StorageGasParameters;
+use aptos_vm_types::storage::io_pricing::IoPricing;
 use move_binary_format::errors::PartialVMResult;
 use std::sync::{Arc, Mutex};
 
@@ -29,8 +29,8 @@ impl<A: GasAlgebra> GasAlgebra for CalibrationAlgebra<A> {
         self.base.vm_gas_params()
     }
 
-    fn storage_gas_params(&self) -> &StorageGasParameters {
-        self.base.storage_gas_params()
+    fn io_pricing(&self) -> &IoPricing {
+        self.base.io_pricing()
     }
 
     fn balance_internal(&self) -> InternalGas {
