+++ aptos-move/aptos-debugger/src/aptos_debugger.rs
@@ -29,7 +29,9 @@ use aptos_vm::{
     AptosVM, VMExecutor,
 };
 use aptos_vm_logging::log_schema::AdapterLogSchema;
-use aptos_vm_types::{change_set::VMChangeSet, output::VMOutput, storage::ChangeSetConfigs};
+use aptos_vm_types::{
+    change_set::VMChangeSet, output::VMOutput, storage::change_set_configs::ChangeSetConfigs,
+};
 use move_binary_format::errors::VMResult;
 use std::{path::Path, sync::Arc};
