+++ aptos-move/aptos-gas-meter/src/algebra.rs
@@ -5,7 +5,7 @@ use crate::traits::GasAlgebra;
 use aptos_gas_algebra::{Fee, FeePerGasUnit, Gas, GasExpression, Octa};
 use aptos_gas_schedule::VMGasParameters;
 use aptos_logger::error;
-use aptos_vm_types::storage::StorageGasParameters;
+use aptos_vm_types::storage::{io_pricing::IoPricing, StorageGasParameters};
 use move_binary_format::errors::{PartialVMError, PartialVMResult};
 use move_core_types::{
     gas_algebra::{InternalGas, InternalGasUnit},
@@ -83,8 +83,8 @@ impl GasAlgebra for StandardGasAlgebra {
         &self.vm_gas_params
     }
 
-    fn storage_gas_params(&self) -> &StorageGasParameters {
-        &self.storage_gas_params
+    fn io_pricing(&self) -> &IoPricing {
+        &self.storage_gas_params.io_pricing
     }
 
     fn balance_internal(&self) -> InternalGas {
